# Conflicts:
#	Market/src/main/java/com/market/config/WebSecurityConfig.java
#	Market/src/main/java/com/market/controller/CosmeticController.java
#	Market/src/main/java/com/market/controller/DrinkController.java
#	Market/src/main/java/com/market/controller/FoodController.java
#	Market/src/main/java/com/market/controller/ProductController.java
#	Market/src/main/java/com/market/controller/SanitaryController.java
#	Market/src/main/java/com/market/controller/UserController.java
#	Market/src/main/java/com/market/entity/Role.java
#	Market/src/main/java/com/market/entity/order/Order.java
#	Market/src/main/java/com/market/entity/order/OrderItem.java
#	Market/src/main/java/com/market/entity/productTypes/Product.java
#	Market/src/main/java/com/market/init/DataBaseInit.java
#	Market/src/main/java/com/market/service/impl/product/CosmeticServiceImpl.java
#	Market/src/main/java/com/market/service/impl/product/ProductServiceImpl.java
#	Market/src/main/java/com/market/service/impl/user/UserServiceImpl.java
#	Market/src/main/java/com/market/service/user/UserService.java
#	Market/src/main/java/com/market/utility/enums/RoleNameEnum.java
#	Market/src/main/resources/templates/fragments/header.html
#	Market/src/main/resources/templates/products.html

+++ Market/src/main/java/com/market/service/impl/product/ProductServiceImpl.java
@@ -1,9 +1,12 @@
 package com.market.service.impl.product;
 
-import com.market.dto.request.ProductCreate;
 import com.market.dto.request.Filter;
+import com.market.dto.request.ProductCreate;
+import com.market.entity.order.OrderItem;
 import com.market.entity.productTypes.Product;
 import com.market.repository.product.ProductRepository;
+import com.market.service.product.SpecificationProductFilter;
+import com.market.service.order.OrderItemService;
 import com.market.service.product.*;
 import com.market.utility.exception.NotFoundException;
 import org.springframework.stereotype.Service;
@@ -19,25 +22,25 @@ public class ProductServiceImpl<P extends Product> implements ProductService {
     private final CosmeticService cosmeticService;
     private final SanitaryService sanitaryService;
     private final SpecificationProductFilter specificationProductFilter;
+    private final OrderItemService orderItemService;
 
     public ProductServiceImpl(ProductRepository productRepository,
                               FoodService foodService,
                               DrinkService drinkService,
                               CosmeticService cosmeticService,
-                              SanitaryService sanitaryService, SpecificationProductFilter specificationProductFilter) {
+                              SanitaryService sanitaryService, SpecificationProductFilter specificationProductFilter, OrderItemService orderItemService) {
         this.productRepository = productRepository;
         this.foodService = foodService;
         this.drinkService = drinkService;
         this.cosmeticService = cosmeticService;
         this.sanitaryService = sanitaryService;
-
         this.specificationProductFilter = specificationProductFilter;
+        this.orderItemService = orderItemService;
     }
 
 
     @Override
     public List<Product> findAll() {
-
         return productRepository.findAll();
     }
 
@@ -46,27 +49,45 @@ public class ProductServiceImpl<P extends Product> implements ProductService {
         return productRepository.findAllAvailable();
     }
 
+    //    SAVE PRODUCT CREATE DTO - OK
     @Override
-    public void saveProduct(ProductCreate productCreate) {
+    public void saveProductCreate(ProductCreate productCreate) {
 
-        switch (productCreate.getType()) {
+        switch (productCreate.getType().name()) {
             case "FOOD":
+
                 foodService.save(productCreate);
                 break;
             case "DRINKS":
+
                 drinkService.save(productCreate);
                 break;
             case "SANITARY":
+
                 sanitaryService.save(productCreate);
                 break;
             case "COSMETICS":
+
                 cosmeticService.save(productCreate);
                 break;
         }
+
     }
 
+    //    DELETE PRODUCT BY ID - OK
     @Override
     public void deleteProductById(Long id) {
+        Product product = productRepository.findById(id).orElse(null);
+
+        if (product == null) {
+            throw new NotFoundException("Продуктът не е намерен");
+        }
+
+        for (OrderItem item : product.getItems()) {
+            item.setProduct(null);
+            item.setQuantity(0);
+            orderItemService.saveItem(item);
+        }
         productRepository.deleteById(id);
     }
 
@@ -76,10 +97,15 @@ public class ProductServiceImpl<P extends Product> implements ProductService {
     }
 
     @Override
+    public void saveProduct(Product product) {
+        productRepository.save(product);
+    }
+
+    @Override
     public void updateProduct(ProductCreate productCreate) {
-        switch (productCreate.getType()) {
+        switch (productCreate.getType().name()) {
             case "FOOD":
-                foodService.updateFood(productCreate);
+                foodService.update(productCreate);
                 break;
             case "DRINKS":
                 drinkService.update(productCreate);
@@ -95,7 +121,9 @@ public class ProductServiceImpl<P extends Product> implements ProductService {
 
 
     public Product findProductById(Long id) {
-        return productRepository.findById(id).orElseThrow(() -> new NotFoundException("product with " + id + " was not found"));
+        return productRepository
+                .findById(id)
+                .orElseThrow(() -> new NotFoundException("product with " + id + " was not found"));
     }
